Web: Files: refactoring: fix after merge
